# research_environment_api/web/static/swagger.json
{"info": {"description": "Health Data Nexus Research Environment API", "title": "Research Environment API", "version": "1.0.0"}, "paths": {"/static/{filename}": {}, "/identity/create": {"post": {"tags": ["identity_management"], "description": "Creates a cloud identity.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/IdentityProvisioningRequest"}}}}, "responses": {"200": {"description": "Returns the created cloud identity.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProvisionedIdentity"}}}}}}}, "/billing/{email}": {"get": {"tags": ["billing_management"], "description": "Lists a user's billing accounts.", "parameters": [{"in": "path", "name": "email", "schema": {"type": "string"}, "required": true}], "responses": {"200": {"description": "Returns a list of billing accounts.", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/BillingAccount"}}}}}}}}, "/billing/share": {"post": {"tags": ["billing_management"], "description": "Shares a billing account.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShareBillingAccountRequest"}}}}, "responses": {"200": {"description": "Returns an empty object.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/None"}}}}}}}, "/billing/revoke_access": {"post": {"tags": ["billing_management"], "description": "Revokes a user's access to a billing account.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RevokeBillingAccountAccessRequest"}}}}, "responses": {"200": {"description": "Returns an empty object.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/None"}}}}}}}, "/workspace/create": {"post": {"tags": ["workspace_management"], "description": "Creates the specified workspace.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceCreationRequest"}}}}, "responses": {"200": {"description": "Returns the ID of the workflow.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceWorkflowIdentifier"}}}}}}}, "/workspace/delete": {"post": {"tags": ["workspace_management"], "description": "Deletes the specified workspace.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceDeletionRequest"}}}}, "responses": {"200": {"description": "Returns the ID of the workflow.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceWorkflowIdentifier"}}}}}}}, "/workspace/{email}": {"get": {"tags": ["workspace_management"], "description": "Lists the active workspaces for a specified user.", "parameters": [{"in": "path", "name": "email", "schema": {"type": "string"}, "required": true}], "responses": {"200": {"description": "Returns a list of workspaces.", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Workspace"}}}}}}}}, "/workbench/create": {"post": {"tags": ["workbench_management"], "description": "Creates a workbench according to the specification.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchCreateRequest"}}}}, "responses": {"200": {"description": "Returns the ID of the workflow.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchWorkflowIdentifier"}}}}}}}, "/workbench/stop": {"put": {"tags": ["workbench_management"], "description": "Stops the specified workbench.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchToggleStateRequest"}}}}, "responses": {"200": {"description": "Returns the ID of the workflow.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchWorkflowIdentifier"}}}}}}}, "/workbench/start": {"put": {"tags": ["workbench_management"], "description": "Starts the specified workbench.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchToggleStateRequest"}}}}, "responses": {"200": {"description": "Returns the ID of the workflow.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchWorkflowIdentifier"}}}}}}}, "/workbench/update": {"put": {"tags": ["workbench_management"], "description": "Updates the specified workbench.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchUpdateRequest"}}}}, "responses": {"200": {"description": "Returns the ID of the workflow.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchWorkflowIdentifier"}}}}}}}, "/workbench/destroy": {"delete": {"tags": ["workbench_management"], "description": "Destroys the specified workbench.", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchDestroyRequest"}}}}, "responses": {"200": {"description": "Returns the ID of the workflow.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchWorkflowIdentifier"}}}}}}}}, "openapi": "3.0.0", "components": {"schemas": {"IdentityProvisioningRequest": {"type": "object", "properties": {"user_name": {"type": "string"}, "recovery_email": {"type": "string"}, "given_name": {"type": "string"}, "family_name": {"type": "string"}, "password": {"type": "string"}}, "required": ["family_name", "given_name", "password", "recovery_email", "user_name"]}, "ProvisionedIdentity": {"type": "object", "properties": {"primary_email": {"type": "string"}, "given_name": {"type": "string"}, "family_name": {"type": "string"}, "recovery_email": {"type": "string"}}, "required": ["family_name", "given_name", "primary_email", "recovery_email"]}, "BillingAccount": {"type": "object", "properties": {"id": {"type": "string"}, "cloud_link": {"type": "string", "format": "url"}, "name": {"type": "string"}, "is_owner": {"type": "boolean"}}, "required": ["cloud_link", "id", "is_owner", "name"]}, "ShareBillingAccountRequest": {"type": "object", "properties": {"user_email": {"type": "string"}, "billing_account_id": {"type": "string"}, "owner_email": {"type": "string"}}, "required": ["billing_account_id", "owner_email", "user_email"]}, "RevokeBillingAccountAccessRequest": {"type": "object", "properties": {"user_email": {"type": "string"}, "billing_account_id": {"type": "string"}, "owner_email": {"type": "string"}}, "required": ["billing_account_id", "owner_email", "user_email"]}, "WorkspaceCreationRequest": {"type": "object", "properties": {"user_email": {"type": "string"}, "billing_account_id": {"type": "string"}, "region": {"enum": ["us-central1", "northamerica-northeast1", "europe-west3", "australia-southeast1"]}}, "required": ["billing_account_id", "region", "user_email"]}, "WorkspaceWorkflowIdentifier": {"type": "object", "properties": {"workflow_id": {"type": "string"}}, "required": ["workflow_id"]}, "WorkspaceDeletionRequest": {"type": "object", "properties": {"user_email": {"type": "string"}, "billing_account_id": {"type": "string"}, "workspace_project_id": {"type": "string"}, "region": {"enum": ["us-central1", "northamerica-northeast1", "europe-west3", "australia-southeast1"]}}, "required": ["billing_account_id", "region", "user_email", "workspace_project_id"]}, "Workbench": {"type": "object", "properties": {"dataset_identifier": {"type": "string"}, "gcp_identifier": {"type": "string"}, "memory": {"type": "number"}, "url": {"type": "string", "format": "url"}, "type": {"enum": ["jupyter", "rstudio"]}, "disk_size": {"type": "integer"}, "cpu": {"type": "number"}, "zone": {"type": "string"}, "status": {"enum": ["running", "stopped"]}, "machine_type": {"enum": ["n1-standard-2", "n1-standard-4", "n1-standard-8", "n1-standard-16"]}}, "required": ["cpu", "dataset_identifier", "disk_size", "gcp_identifier", "machine_type", "memory", "status", "type", "url"]}, "BillingInfo": {"type": "object", "properties": {"billing_account_id": {"type": "string"}, "billing_enabled": {"type": "boolean"}}, "required": ["billing_account_id", "billing_enabled"]}, "Workspace": {"type": "object", "properties": {"gcp_project_id": {"type": "string"}, "workbenches": {"type": "array", "items": {"$ref": "#/components/schemas/Workbench"}}, "billing_info": {"$ref": "#/components/schemas/BillingInfo"}, "region": {"enum": ["us-central1", "northamerica-northeast1", "europe-west3", "australia-southeast1"]}}, "required": ["billing_info", "gcp_project_id", "region"]}, "WorkbenchCreateRequest": {"type": "object", "properties": {"dataset_identifier": {"type": "string"}, "bucket_name": {"type": "string"}, "workspace_project_id": {"type": "string"}, "workbench_type": {"enum": ["jupyter", "rstudio"]}, "disk_size": {"type": "integer"}, "user_email": {"type": "string"}, "gpu_accelerator_type": {"nullable": true, "enum": ["NVIDIA_TESLA_T4"]}, "machine_type": {"enum": ["n1-standard-2", "n1-standard-4", "n1-standard-8", "n1-standard-16"]}}, "required": ["bucket_name", "dataset_identifier", "disk_size", "machine_type", "user_email", "workbench_type", "workspace_project_id"]}, "WorkbenchWorkflowIdentifier": {"type": "object", "properties": {"workflow_id": {"type": "string"}}, "required": ["workflow_id"]}, "WorkbenchToggleStateRequest": {"type": "object", "properties": {"user_email": {"type": "string"}, "workspace_project_id": {"type": "string"}, "workbench_type": {"enum": ["jupyter", "rstudio"]}, "workbench_resource_id": {"type": "string"}}, "required": ["user_email", "workbench_resource_id", "workbench_type", "workspace_project_id"]}, "WorkbenchUpdateRequest": {"type": "object", "properties": {"workspace_project_id": {"type": "string"}, "workbench_type": {"enum": ["jupyter", "rstudio"]}, "workbench_resource_id": {"type": "string"}, "user_email": {"type": "string"}, "machine_type": {"enum": ["n1-standard-2", "n1-standard-4", "n1-standard-8", "n1-standard-16"]}}, "required": ["machine_type", "user_email", "workbench_resource_id", "workbench_type", "workspace_project_id"]}, "WorkbenchDestroyRequest": {"type": "object", "properties": {"user_email": {"type": "string"}, "workspace_project_id": {"type": "string"}, "workbench_type": {"enum": ["jupyter", "rstudio"]}, "workbench_resource_id": {"type": "string"}}, "required": ["user_email", "workbench_resource_id", "workbench_type", "workspace_project_id"]}}}}